+++ Aula-1/Ex6.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 
 int Ex6(void) {
-  int n1, n2, n3, n4, media;
+  float n1, n2, n3, n4, media;
 
   printf("Digite a nota 1\n");
-  scanf("%d", &n1);
+  scanf("%f", &n1);
   printf("Digite a nota 2\n");
-  scanf("%d", &n2);
+  scanf("%f", &n2);
   printf("Digite a nota 3\n");
-  scanf("%d", &n3);
+  scanf("%f", &n3);
   printf("Digite a nota 4\n");
-  scanf("%d", &n4);
+  scanf("%f", &n4);
 
   media = (n1 + n2 + n3 + n4) / 4;
 
-  printf("A Média é: %d\n", media);
+  printf("A Média é: %f\n", media);
 
   return 0;
 }
